spark/requirements.txt
pyspark==3.5.3
pyyaml
